pyproject.toml
[tool.poetry]
name = "healthy-py"
version = "0.1.0"
description = ""
authors = ["Nathan Jones <nathanjones4323@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.3"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
metabase-api = "^3.5"
pandas = "^2.3.0"
sqlalchemy = "^2.0.41"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
